Draft Generate report impl and ResidentCharge persistence logic

+++ src/main/java/com/jbn/block/maintenance/residence/details/dto/ResidentDetailDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ResidentDetailDto {
 
-    private Long residentId;
+    private Integer residentId;
     private String fullName;
     private Integer age;
     private Long doorNo;
